# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/interviews": {
      "post": {
        "operationId": "InterviewController_create",
        "summary": "Create interview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewWithCandidateLinkResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      },
      "get": {
        "operationId": "InterviewController_findAll",
        "summary": "List interviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
...
    "/interviews/{id}/questions/{questionIndex}/media": {
      "get": {
        "operationId": "RecruiterMediaController_getAnswerMedia",
        "summary": "Get signed media URLs for interview answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionIndex",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewAnswerMediaResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "interviews"
        ]
      }
    },
...
    "/take/{id}": {
      "get": {
        "operationId": "TakeController_getInterview",
        "summary": "Get candidate interview state",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TakeInterviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "take"
        ]
      }
    },
    "/take/{id}/answer": {
      "post": {
        "operationId": "TakeController_submitAnswer",
        "summary": "Submit candidate answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitTakeAnswerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "take"
        ]
      }
    },
    "/take/{id}/answer/progress": {
      "post": {
        "operationId": "TakeController_saveAnswerProgress",
        "summary": "Save candidate answer progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAnswerProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveTakeAnswerProgressResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "take"
        ]
      }
    },
    "/take/{id}/questions/{questionIndex}/validate": {
      "post": {
        "operationId": "TakeController_startAnswerValidation",
        "summary": "Start candidate answer validation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionIndex",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTakeAnswerValidationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "take"
        ]
      }
    },
...
  },
  "info": {
    "title": "Interview App API",
    "description": "HTTP API contract for Interview App backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "focus": {
            "type": "string"
          },
          "outputLanguage": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "questionText": {
            "type": "string"
          },
          "followUpQuestions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expectedConcepts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionExpectedConceptDto"
            }
          },
          "redFlags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionRedFlagDto"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "weight": {
            "type": "number"
          },
          "sampleGoodAnswer": {
            "type": "string"
          },
          "minimumPassScore": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "outputLanguage",
          "questionText",
          "followUpQuestions",
          "expectedConcepts",
          "redFlags",
          "difficulty",
          "weight",
          "minimumPassScore",
          "tags",
          "metadata",
          "createdAt",
          "updatedAt",
          "deleted"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Validation failed"
          },
          "path": {
            "type": "string",
            "example": "/questions/invalid-id"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "CreateInterviewDto": {
        "type": "object",
        "properties": {
          "candidateName": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "questionIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "candidateName",
          "position",
          "questionIds"
        ]
      },
...
      "AnswerTranscriptDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "generatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isFinal": {
            "type": "boolean"
          }
        }
      },
...
      "AnswerDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "questionId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "recording",
              "submitted"
            ]
          },
          "mediaKey": {
            "type": "string"
          },
          "screenMediaKey": {
            "type": "string"
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          },
          "durationSeconds": {
            "type": "number"
          },
          "retakeCount": {
            "type": "number"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "camera": {
            "$ref": "#/components/schemas/MediaArtifactDto"
          },
          "screen": {
            "$ref": "#/components/schemas/MediaArtifactDto"
          },
          "behaviorSignals": {
            "$ref": "#/components/schemas/AnswerBehaviorSignalsDto"
          },
          "selectedVersionNumber": {
            "type": "number"
          },
          "transcript": {
            "$ref": "#/components/schemas/AnswerTranscriptDto"
          },
          "evaluation": {
            "$ref": "#/components/schemas/AnswerEvaluationDto"
          },
          "validation": {
            "$ref": "#/components/schemas/AnswerValidationDto"
          }
        },
        "required": [
          "questionIndex",
          "questionId",
          "status",
          "mediaKey",
          "uploadedAt"
        ]
      },
...
      "InterviewWithCandidateLinkResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateName": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "processing",
              "completed",
              "failed"
            ]
          },
          "result": {
            "$ref": "#/components/schemas/InterviewResultResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "candidateLink": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidateName",
          "position",
          "questions",
          "answers",
          "status",
          "createdAt",
          "updatedAt",
          "candidateLink"
        ]
      },
      "InterviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateName": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          },
          "answers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnswerDto"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "in_progress",
              "processing",
              "completed",
              "failed"
            ]
          },
          "result": {
            "$ref": "#/components/schemas/InterviewResultResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "candidateName",
          "position",
          "questions",
          "answers",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "InterviewAnswerMediaResponseDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "cameraUrl": {
            "type": "string"
          },
          "screenUrl": {
            "type": "string"
          }
        },
        "required": [
          "questionIndex"
        ]
      },
...
      "CandidateQuestionViewDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      },
...
      "TakeInterviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "candidateName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalQuestions": {
            "type": "number"
          },
          "currentQuestion": {
            "$ref": "#/components/schemas/CandidateQuestionViewDto"
          },
          "currentQuestionIndex": {
            "type": "number"
          },
          "currentAnswerMeta": {
            "$ref": "#/components/schemas/CurrentAnswerMetaDto"
          },
          "completed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "position",
          "candidateName",
          "status",
          "totalQuestions",
          "currentQuestionIndex",
          "completed"
        ]
      },
...
      "SubmitAnswerDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "versionNumber": {
            "type": "number"
          },
          "submitAnswer": {
            "type": "boolean"
          },
          "mediaKey": {
            "type": "string"
          },
          "screenMediaKey": {
            "type": "string"
          },
          "durationSeconds": {
            "type": "number"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "cameraFileSizeBytes": {
            "type": "number"
          },
          "screenFileSizeBytes": {
            "type": "number"
          },
          "behaviorSignals": {
            "$ref": "#/components/schemas/BehaviorSignalsDto"
          },
          "behaviorEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BehaviorEventDto"
            }
          },
          "clientTranscript": {
            "$ref": "#/components/schemas/ClientTranscriptDto"
          }
        },
        "required": [
          "questionIndex",
          "versionNumber",
          "submitAnswer",
          "mediaKey",
          "durationSeconds",
          "startedAt",
          "submittedAt",
          "behaviorSignals"
        ]
      },
      "SubmitTakeAnswerResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "answeredCount": {
            "type": "number"
          },
          "totalQuestions": {
            "type": "number"
          },
          "completed": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "answeredCount",
          "totalQuestions",
          "completed"
        ]
      },
      "SaveAnswerProgressDto": {
        "type": "object",
        "properties": {
          "questionIndex": {
            "type": "number"
          },
          "versionNumber": {
            "type": "number"
          },
          "mediaKey": {
            "type": "string"
          },
          "screenMediaKey": {
            "type": "string"
          },
          "durationSeconds": {
            "type": "number"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "cameraFileSizeBytes": {
            "type": "number"
          },
          "screenFileSizeBytes": {
            "type": "number"
          },
          "behaviorSignals": {
            "$ref": "#/components/schemas/BehaviorSignalsDto"
          },
          "behaviorEvents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BehaviorEventDto"
            }
          },
          "clientTranscript": {
            "$ref": "#/components/schemas/ClientTranscriptDto"
          }
        },
        "required": [
          "questionIndex",
          "versionNumber",
          "mediaKey",
          "behaviorSignals"
        ]
      },
      "SaveTakeAnswerProgressResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "recording",
              "submitted"
            ]
          },
          "versionCount": {
            "type": "number"
          },
          "selectedVersionNumber": {
            "type": "number"
          }
        },
        "required": [
          "ok",
          "status",
          "versionCount",
          "selectedVersionNumber"
        ]
      },
      "StartTakeAnswerValidationResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "queued",
              "processing",
              "completed",
              "failed"
            ]
          },
          "questionIndex": {
            "type": "number"
          },
          "sourceVersionNumber": {
            "type": "number"
          },
          "reused": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "status",
          "questionIndex",
          "sourceVersionNumber",
          "reused"
        ]
      },
...
    }
  }
}