# src/upload/recruiter-media.controller.ts
import {
  Controller,
  Get,
  NotFoundException,
  Param,
  ParseIntPipe,
  UseGuards,
} from '@nestjs/common';
import {
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiTags,
  ApiUnauthorizedResponse,
} from '@nestjs/swagger';
import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
import { RolesGuard } from '../auth/guards/roles.guard';
import { Roles } from '../auth/decorators/roles.decorator';
import { InterviewService } from '../interview/interview.service';
import { UploadService } from './upload.service';
import { InterviewAnswerMediaResponseDto } from './dto/upload.responses.dto';
import { ApiErrorResponseDto } from '../common/dto/api-error.response.dto';

@ApiTags('interviews')
@Controller('interviews')
@UseGuards(JwtAuthGuard, RolesGuard)
@Roles('super_admin', 'admin', 'hr')
export class RecruiterMediaController {
  constructor(
    private readonly interviewService: InterviewService,
    private readonly uploadService: UploadService,
  ) {}

  @Get(':id/questions/:questionIndex/media')
  @ApiOperation({ summary: 'Get signed media URLs for interview answer' })
  @ApiParam({ name: 'id' })
  @ApiParam({ name: 'questionIndex' })
  @ApiOkResponse({ type: InterviewAnswerMediaResponseDto })
  @ApiUnauthorizedResponse({ type: ApiErrorResponseDto })
  @ApiNotFoundResponse({ type: ApiErrorResponseDto })
  async getAnswerMedia(
    @Param('id') id: string,
    @Param('questionIndex', ParseIntPipe) questionIndex: number,
  ) {
    const interview = await this.interviewService.findOne(id);
    const answer = interview.answers.find(
      (item) => item.questionIndex === questionIndex,
    );

    if (!answer) {
      throw new NotFoundException(
        `Answer for question ${questionIndex} is not available`,
      );
    }

    const [camera, screen] = await Promise.all([
      answer.mediaKey
        ? this.uploadService.generateDownloadUrl(
            id,
            questionIndex,
            answer.mediaKey,
          )
        : Promise.resolve(null),
      answer.screenMediaKey
        ? this.uploadService.generateDownloadUrl(
            id,
            questionIndex,
            answer.screenMediaKey,
          )
        : Promise.resolve(null),
    ]);

    return {
      questionIndex,
      cameraUrl: camera?.downloadUrl,
      screenUrl: screen?.downloadUrl,
    };
  }
}